Refaactored some input codes in Player. Renamed package name from "saving" to "data".
