+++ game/src/main/java/data/PlayerInfo.java
@@ -8,7 +8,7 @@
  * THE CREATOR IS NOT LIABLE FOR ANY DAMAGES DONE. FOLLOW LOCAL LAWS, BE RESPECTFUL, AND HAVE A GOOD DAY!
  * */
 
-package saving;
+package data;
 
 import java.io.IOException;
 import java.io.RandomAccessFile;
