+++ game/src/main/java/main/Game.java
@@ -15,6 +15,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import abstracts.SubMenu;
+import data.GameSave;
 import entity.Player;
 import item.ActionItem;
 import level.OverWorld;
@@ -25,7 +26,6 @@ import menu.Inventory;
 import menu.MainMenu;
 import resources.Art;
 import resources.Mod;
-import saving.GameSave;
 import screen.Scene;
 
 public class Game {
