+++ game/src/main/java/main/SaveDataManager.java
@@ -6,11 +6,11 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 
 import abstracts.SubMenu;
+import data.GameSave;
 import dialogue.Dialogue;
 import entity.Player;
 import level.WorldConstants;
 import main.StateManager.GameState;
-import saving.GameSave;
 import screen.Scene;
 import utility.DialogueBuilder;
